requirements.txt
streamlit
streamlit-authenticator
pandas
pypdf
PyYAML